updating files

+++ week02/bmi.py
@@ -2,8 +2,8 @@
 # author: Susan Henry
 
 #input persons weight in kilograms and height in cms
-weight = float (input ('Enter weight in kg:'))
-height = float (input ('Enter height in cms:'))
+weight = int (input ('Enter weight in kg:'))
+height = int (input ('Enter height in cms:'))
 
 #calcution for bmi 
 bmi = round (weight / (height/100 * height/100), 2)
